+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0 - 2021-07-03
+### Added
+- Added full class analysis.
+- Added function analysis.
+
+### Refactor
+- Refactored reader and compiler to be static.
+- Renamed classes in registry to targets and renamed registry to DecoratedRegistry.
+
 ## 1.1.0 - 2021-07-03
 ### Added
 - Method attribute analysis.
@@ -13,5 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial implementation of the package.
 
-[1.1.0 > Unreleased](https://github.com/grizz-it/metadata/compare/1.1.0...HEAD)
+[2.0.0 > Unreleased](https://github.com/grizz-it/metadata/compare/2.0.0...HEAD)
+[1.1.0 > 2.0.0](https://github.com/grizz-it/metadata/compare/1.1.0...2.0.0)
 [1.0.0 > 1.1.0](https://github.com/grizz-it/metadata/compare/1.0.0...1.1.0)